oceanexplorer

This package allows easy exploration of the world ocean atlas of the US agency National Oceanic and Atmospheric Administration (NOAA). It includes functions to extract NetCDF data from the repository and code to visualize several physical and chemical parameters of the ocean. A Shiny app further allows interactive exploration of the data.

See also

Author

Maintainer: Martin Schobben schobbenmartin@gmail.com (ORCID) [copyright holder]

Other contributors:

Peter Bijl (ORCID) [contributor]